+++ packages/commonwealth/client/scripts/views/pages/Communities/Communities.tsx
@@ -1,53 +1,29 @@
-import { ExtendedCommunity } from '@hicommonwealth/schemas';
-import { ChainNetwork, CommunityType } from '@hicommonwealth/shared';
-import clsx from 'clsx';
 import { findDenominationString } from 'helpers/findDenomination';
 import useBrowserWindow from 'hooks/useBrowserWindow';
 import { useFlag } from 'hooks/useFlag';
 import { useCommonNavigate } from 'navigation/helpers';
-import React, { useCallback, useMemo, useRef, useState } from 'react';
+import React, { useRef, useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
-import { useFetchCommunitiesQuery } from 'state/api/communities';
-import { useFetchTagsQuery } from 'state/api/tags';
 import { useManageCommunityStakeModalStore } from 'state/ui/modals';
 import CWPageLayout from 'views/components/component_kit/new_designs/CWPageLayout';
-import { z } from 'zod';
 import { useFetchTokenUsdRateQuery } from '../../../state/api/communityStake/index';
 import { useFetchGlobalActivityQuery } from '../../../state/api/feeds/fetchUserActivity';
 import { trpc } from '../../../utils/trpcClient';
 import { CWText } from '../../components/component_kit/cw_text';
-import { CWButton } from '../../components/component_kit/new_designs/CWButton';
 import { CWModal } from '../../components/component_kit/new_designs/CWModal';
 import CWTab from '../../components/component_kit/new_designs/CWTabs/CWTab';
 import CWTabsRow from '../../components/component_kit/new_designs/CWTabs/CWTabsRow';
-import { CWTag } from '../../components/component_kit/new_designs/CWTag';
 import { Feed } from '../../components/feed';
 import CreateCommunityButton from '../../components/sidebar/CreateCommunityButton';
 import ManageCommunityStakeModal from '../../modals/ManageCommunityStakeModal/ManageCommunityStakeModal';
 import XPTable from '../Leaderboard/XPTable/XPTable';
 import AllTabContent from './AllTabContent';
 import './Communities.scss';
-import CommunitiesTabContent from './CommunitiesTabContent';
+import CommunitiesList from './CommunitiesList';
 import ExploreContestList from './ExploreContestList';
-import {
-  CommunityFilters,
-  CommunitySortDirections,
-  CommunitySortOptions,
-  FiltersDrawer,
-  communityChains,
-  communitySortOptionsLabelToKeysMap,
-  sortOrderLabelsToDirectionsMap,
-} from './FiltersDrawer';
 import IdeaLaunchpad from './IdeaLaunchpad';
 import QuestList from './QuestList';
 import TokensList from './TokensList';
-import { getCommunityCountsString } from './helpers';
-
-type ExtendedCommunityType = z.infer<typeof ExtendedCommunity>;
-type ExtendedCommunitySliceType = [
-  ExtendedCommunityType,
-  ExtendedCommunityType,
-];
 
 const CommunitiesPage = () => {
   const containerRef = useRef();
@@ -75,79 +51,11 @@ const CommunitiesPage = () => {
     modeOfManageCommunityStakeModal,
   } = useManageCommunityStakeModalStore();
 
-  const [filters, setFilters] = useState<CommunityFilters>({
-    withCommunityEcosystem: undefined,
-    withStakeEnabled: undefined,
-    withTagsIds: undefined,
-    withCommunitySortBy: CommunitySortOptions.MemberCount,
-    withCommunitySortOrder: CommunitySortDirections.Descending,
-    withCommunityType: undefined,
-    withEcosystemChainId: undefined,
-    withNetwork: undefined,
-  });
-
   const [selectedCommunityId, setSelectedCommunityId] = useState<string>();
 
   const oneDayAgo = useRef(new Date().getTime() - 24 * 60 * 60 * 1000);
 
-  const { data: tags, isLoading: isLoadingTags } = useFetchTagsQuery();
-
   const { isWindowSmallInclusive } = useBrowserWindow({});
-
-  const {
-    data: communities,
-    fetchNextPage: fetchMoreCommunitiesOriginal,
-    hasNextPage,
-    isInitialLoading: isInitialCommunitiesLoading,
-  } = useFetchCommunitiesQuery({
-    limit: 50,
-    include_node_info: true,
-    order_by: (() => {
-      if (
-        filters.withCommunitySortBy &&
-        [
-          CommunitySortOptions.MemberCount,
-          CommunitySortOptions.ThreadCount,
-          CommunitySortOptions.MostRecent,
-        ].includes(filters.withCommunitySortBy)
-      ) {
-        return (
-          (communitySortOptionsLabelToKeysMap[
-            filters.withCommunitySortBy
-            // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          ] as any) || 'lifetime_thread_count'
-        );
-      }
-
-      return 'lifetime_thread_count';
-    })(),
-    order_direction:
-      sortOrderLabelsToDirectionsMap[filters.withCommunitySortOrder || ''] ||
-      'DESC',
-    eth_chain_id:
-      typeof filters.withEcosystemChainId === 'number'
-        ? filters.withEcosystemChainId
-        : undefined,
-    cosmos_chain_id:
-      typeof filters.withEcosystemChainId === 'string'
-        ? filters.withEcosystemChainId
-        : undefined,
-    base: filters.withCommunityEcosystem || undefined,
-    network: filters.withNetwork
-      ? ChainNetwork[filters.withNetwork]
-      : undefined,
-    stake_enabled: filters.withStakeEnabled,
-    cursor: 1,
-    tag_ids: filters.withTagsIds,
-    community_type: filters.withCommunityType
-      ? CommunityType[filters.withCommunityType]
-      : undefined,
-  });
-
-  // Wrap fetchMoreCommunities to return Promise<void>
-  const fetchMoreCommunities = useCallback(async () => {
-    await fetchMoreCommunitiesOriginal();
-  }, [fetchMoreCommunitiesOriginal]);
 
   const { data: historicalPrices, isLoading: isLoadingHistoricalPrices } =
     trpc.community.getStakeHistoricalPrice.useQuery({
@@ -160,88 +68,7 @@ const CommunitiesPage = () => {
     });
   const ethUsdRate = ethUsdRateData?.data?.data?.amount;
 
-  const [isFilterDrawerOpen, setIsFilterDrawerOpen] = useState(false);
-
-  const isLoading =
-    isLoadingTags ||
-    isInitialCommunitiesLoading ||
-    isLoadingHistoricalPrices ||
-    isLoadingEthUsdRate;
-
-  const communitiesList = useMemo(() => {
-    const flatList = (communities?.pages || []).flatMap((page) => page.results);
-
-    const SLICE_SIZE = 2;
-    const twoCommunitiesPerEntry: ExtendedCommunitySliceType[] = [];
-
-    for (let i = 0; i < flatList.length; i += SLICE_SIZE) {
-      twoCommunitiesPerEntry.push(
-        flatList.slice(i, i + SLICE_SIZE) as ExtendedCommunitySliceType,
-      );
-    }
-
-    return twoCommunitiesPerEntry;
-  }, [communities?.pages]);
-
-  const removeStakeFilter = () => {
-    setFilters({
-      ...filters,
-      withStakeEnabled: false,
-    });
-  };
-
-  const removeTagFilter = (tagId: number) => {
-    setFilters({
-      ...filters,
-      withTagsIds: [...(filters.withTagsIds || [])].filter(
-        (id) => tagId !== id,
-      ),
-    });
-  };
-
-  const removeCommunityEcosystemFilter = () => {
-    setFilters({
-      ...filters,
-      withCommunityEcosystem: undefined,
-    });
-  };
-
-  const removeEcosystemChainIdFilter = () => {
-    setFilters({
-      ...filters,
-      withEcosystemChainId: undefined,
-    });
-  };
-
-  const removeChainNetworkFilter = () => {
-    setFilters({
-      ...filters,
-      withNetwork: undefined,
-    });
-  };
-
-  const removeCommunityTypeFilter = () => {
-    setFilters({
-      ...filters,
-      withCommunityType: undefined,
-    });
-  };
-
-  const removeCommunitySortByFilter = () => {
-    setFilters({
-      ...filters,
-      withCommunitySortBy: undefined,
-      withCommunitySortOrder: undefined,
-    });
-  };
-
-  const communitiesCount = (
-    <CWText type="b2" className="communities-count">
-      {!isLoading && communities?.pages?.[0]?.totalResults
-        ? getCommunityCountsString(communities?.pages?.[0]?.totalResults)
-        : 'No communities found'}
-    </CWText>
-  );
+  const isLoading = isLoadingHistoricalPrices || isLoadingEthUsdRate;
 
   // Function to handle tab switching
   const handleTabClick = (tabValue: string) => {
@@ -249,6 +76,8 @@ const CommunitiesPage = () => {
     params.set('tab', tabValue);
     navigate(`/explore?${params.toString()}`);
   };
+
+  const communitiesCount = <></>; // TODO: fix this
 
   return (
     // @ts-expect-error <StrictNullChecks/>
@@ -289,9 +118,7 @@ const CommunitiesPage = () => {
 
         {/* Conditionally render content based on active tab */}
         {launchpadEnabled
-          ? activeTab === 'tokens' && (
-              <TokensList filters={filters} hideHeader />
-            )
+          ? activeTab === 'tokens' && <TokensList hideHeader />
           : null}
         {questsEnabled
           ? activeTab === 'quests' && <QuestList hideHeader />
@@ -318,7 +145,7 @@ const CommunitiesPage = () => {
           <>
             {/* Communities section */}
             <div className="section-container">
-              <AllTabContent containerRef={containerRef} filters={filters} />
+              <AllTabContent containerRef={containerRef} />
             </div>
           </>
         )}
@@ -326,100 +153,12 @@ const CommunitiesPage = () => {
         {/* Communities Tab Content */}
         {activeTab === 'communities' && (
           <>
-            <div
-              className={clsx('filters', {
-                hasAppliedFilter:
-                  Object.values(filters).filter(Boolean).length === 1
-                    ? !filters.withCommunitySortOrder
-                    : Object.values(filters).filter(Boolean).length > 0,
-              })}
-            >
-              <CWButton
-                label="Filters"
-                iconRight="funnelSimple"
-                buttonType="secondary"
-                onClick={() => setIsFilterDrawerOpen((isOpen) => !isOpen)}
-              />
-              {filters.withCommunitySortBy && (
-                <CWTag
-                  label={`${filters.withCommunitySortBy}${
-                    filters.withCommunitySortOrder &&
-                    filters.withCommunitySortBy !==
-                      CommunitySortOptions.MostRecent
-                      ? ` : ${filters.withCommunitySortOrder}`
-                      : ''
-                  }
-                  `}
-                  type="filter"
-                  onCloseClick={removeCommunitySortByFilter}
-                />
-              )}
-              {filters.withCommunityType && (
-                <CWTag
-                  label={filters.withCommunityType}
-                  type="filter"
-                  onCloseClick={removeCommunityTypeFilter}
-                />
-              )}
-              {filters.withNetwork && (
-                <CWTag
-                  label={filters.withNetwork}
-                  type="filter"
-                  onCloseClick={removeChainNetworkFilter}
-                />
-              )}
-              {filters.withCommunityEcosystem && (
-                <CWTag
-                  label={filters.withCommunityEcosystem}
-                  type="filter"
-                  onCloseClick={removeCommunityEcosystemFilter}
-                />
-              )}
-              {filters.withEcosystemChainId && (
-                <CWTag
-                  label={
-                    Object.entries(communityChains).find(
-                      ([_, v]) => filters.withEcosystemChainId === v,
-                    )?.[0] as string
-                  }
-                  type="filter"
-                  onCloseClick={removeEcosystemChainIdFilter}
-                />
-              )}
-              {filters.withStakeEnabled && (
-                <CWTag
-                  label="Stake"
-                  type="filter"
-                  onCloseClick={removeStakeFilter}
-                />
-              )}
-              {filters.withTagsIds &&
-                filters.withTagsIds.map((id) => (
-                  <CWTag
-                    key={id}
-                    type="filter"
-                    label={(tags || []).find((t) => t.id === id)?.name || ''}
-                    onCloseClick={() => removeTagFilter(id)}
-                  />
-                ))}
-              <FiltersDrawer
-                isOpen={isFilterDrawerOpen}
-                onClose={() => setIsFilterDrawerOpen(false)}
-                filters={filters}
-                onFiltersChange={(newFilters) => setFilters(newFilters)}
-              />
-            </div>
-            <CommunitiesTabContent
+            <CommunitiesList
               isLoading={isLoading}
-              isInitialCommunitiesLoading={isInitialCommunitiesLoading}
-              communitiesList={communitiesList}
               containerRef={containerRef}
-              filters={filters}
               historicalPrices={historicalPrices}
               ethUsdRate={Number(ethUsdRate)}
               setSelectedCommunityId={setSelectedCommunityId}
-              hasNextPage={hasNextPage}
-              fetchMoreCommunities={fetchMoreCommunities}
             />
           </>
         )}
