+++ packages/commonwealth/client/scripts/views/pages/Communities/CommunitiesList/FiltersDrawer/FiltersDrawer.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/shared.scss';
+@import '../../../../../styles/shared.scss';
 
 .FiltersDrawer {
   .Drawer.filter-drawer {
