+++ packages/commonwealth/client/scripts/views/pages/Communities/CommunitiesList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CommunitiesList';
